requirements.txt
# arXiv API
arxiv

# HTTP requests
requests

# PDF Processing
markitdown[all]

# Utilities
python-dotenv

# State Machine
python-statemachine

# LLM Providers
anthropic
openai